refactor(nav): DRY active link logic and Escape key into shared header-nav module
